Nearly finished all necessary parts for version 1.0. Fitting of Center works now. Savin is still an issue though.

+++ libs/hist/mpa1dhist.cpp
@@ -21,7 +21,7 @@ void Mpa1dHist::setSize(int size){
     mEnergyScale.resize(mSize);
 }
 
-void Mpa1dHist::setBinContent(int nBin,int counts){
+void Mpa1dHist::setBinContent(int nBin,double counts){
     mRawHist(nBin) = counts;
 }
