+++ libs/hist/mpa1dhist.h
@@ -11,12 +11,12 @@ public:
     Mpa1dHist(QString name,int size);
     ~Mpa1dHist();
     void setSize(int nBins);
-    void setBinContent(int nBin,int counts);
+    void setBinContent(int nBin,double counts);
     void setCalibration(float channelToEnergy,float peakEnergy,float peakPosition);
     QString mName;
     int mSize;
-    Eigen::VectorXf mRawHist;
-    Eigen::VectorXf mEnergyScale;
+    Eigen::VectorXd mRawHist;
+    Eigen::VectorXd mEnergyScale;
 };
 
 #endif // MPA1DHIST_H
